requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
databases==0.9.0
pydantic==2.10.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
alembic==1.14.0
email-validator==2.2.0
pytest==8.3.4
httpx==0.28.1
Faker==18.9.0